Discontinued detailed change log.
There is no need for a detailed change log file now that the project is
tracked with Git, a better log can be retrieved automatically from the
version control system instead. Replaced the change log file with a simpler
file just giving a very top-level description of the changes for each
version.